Refactor out observable implementation
